Add tests for new fixer
